requirements.txt
WMI==1.5.1
py_cpuinfo==8.0.0
tabulate==0.8.9
psutil==5.8.0
pandas==1.3.3
GPUtil==1.4.0
Pillow==8.3.2
